@@ -11,27 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.0",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/forms": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/router": "^17.0.0",
+    "@angular/animations": "^17.2.0-rc.1",
+    "@angular/common": "^17.2.0-rc.1",
+    "@angular/compiler": "^17.2.0-rc.1",
+    "@angular/core": "^17.2.0-rc.1",
+    "@angular/forms": "^17.2.0-rc.1",
+    "@angular/platform-browser": "^17.2.0-rc.1",
+    "@angular/platform-browser-dynamic": "^17.2.0-rc.1",
+    "@angular/router": "^17.2.0-rc.1",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.10",
+    "@angular-devkit/build-angular": "^17.2.0-rc.0",
     "@angular-eslint/builder": "17.2.1",
     "@angular-eslint/eslint-plugin": "17.2.1",
     "@angular-eslint/eslint-plugin-template": "17.2.1",
     "@angular-eslint/schematics": "17.2.1",
     "@angular-eslint/template-parser": "17.2.1",
-    "@angular/cli": "^17.0.10",
-    "@angular/compiler-cli": "^17.0.0",
+    "@angular/cli": "^17.2.0-rc.0",
+    "@angular/compiler-cli": "^17.2.0-rc.1",
     "@types/jasmine": "~5.1.0",
     "@typescript-eslint/eslint-plugin": "6.19.0",
     "@typescript-eslint/parser": "6.19.0",
