@@ -1,9 +1,11 @@
 import {
 	ChangeDetectionStrategy,
 	Component,
-	Input,
 	OnInit,
 	inject,
+	input,
+	Signal,
+	computed,
 } from '@angular/core';
 import { TripComponent } from '../trip/trip.component';
 import { TripType } from '@src/app/shared/model/trip';
@@ -21,17 +23,18 @@ import { StorageService } from '@src/app/shared/services/storage.service';
 	changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TripsComponent implements OnInit {
-	@Input() filterPhraze: string = '';
+	filterPhraze = input('');
 	weatherService = inject(WeatherService);
 	storageService = inject(StorageService);
-	trips!: TripType[];
-
-	ngOnInit(): void {
-		this.trips = this.storageService
+	trips: Signal<TripType[]> = computed(() =>
+		this.storageService
 			.getTrips()
 			.filter(trip =>
-				trip.city.toLowerCase().includes(this.filterPhraze.toLowerCase())
-			);
+				trip.city.toLowerCase().includes(this.filterPhraze().toLowerCase())
+			)
+	);
+
+	ngOnInit(): void {
 		this.weatherService.getWeather(
 			CITIES.NewYorkCity.requestKey,
 			'2024-02-19',
